wallpapers: cast the mode in the conditional, which g++ rejects and clang does not
`cond ? uint8_t : SleepScreenMode` is "enumerated and non-enumerated type in
conditional expression" under -Wextra, which is -Werror in the gate's
cross-compiler pass. Clang says nothing, so the local suites were green and
`gcc FAILED under g++-16 in host-tests/ui` and `in host-tests/wallcaption`
was the first sign. The ci-gcc-clang-gap memory, again.

Verified with CXX=g++-16 on wallpapers (479 checks), wallcaption (1045) and
ui (82824), all green.

Docs: none affected.

+++ src/apps_local/wallpapers/WallpapersCore.cpp
@@ -178,7 +178,10 @@ SleepChoice choiceForSetWallpaper(const uint8_t sleepScreenMode, const bool quic
   // classified by the rules and not by this function's memory of them.
   const bool modeAlreadyShowsIt = drawsPinnedSleep(sleepScreenMode, /*quickResumeAfterTimeout=*/false,
                                                    /*fromTimeout=*/true, /*fromReader=*/false);
-  choice.sleepScreenMode = modeAlreadyShowsIt ? sleepScreenMode : kSleepCustom;
+  // The cast is load-bearing under g++ -Wextra: `cond ? uint8_t : SleepScreenMode`
+  // is "enumerated and non-enumerated type in conditional expression", which is
+  // an error in CI and silent under clang (see the ci-gcc-clang-gap memory).
+  choice.sleepScreenMode = modeAlreadyShowsIt ? sleepScreenMode : static_cast<uint8_t>(kSleepCustom);
   choice.tookOverMode = !modeAlreadyShowsIt && sleepScreenMode != kSleepCustom;
 
   // The timeout flag always comes off. It is not a preference about sleep
